# doc/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tiny OIDC JSON APIs",
    "version": "1.0.0",
    "description": "Interaction, Self-service and Admin APIs of Tiny OIDC. The OIDC protocol endpoints are described by the discovery document."
  },
  "components": {
    "securitySchemes": {
      "adminToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "at+jwt",
        "description": "An access token with scope admin and the issuer in aud (§9.1)"
      },
      "accountToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "at+jwt",
        "description": "A user's access token with scope account and the issuer in aud (§8)"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "error_description": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "error_description",
          "request_id"
        ]
      },
...
      "MeProfile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "email_verified": {
            "type": "boolean"
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "groups": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "created_at": {
            "type": "integer"
          },
          "updated_at": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "email",
          "email_verified",
          "display_name",
          "groups",
          "created_at",
          "updated_at"
        ]
      },
      "MeProfilePatch": {
        "type": "object",
        "properties": {
          "display_name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "email": {
            "type": "string",
            "minLength": 3,
            "maxLength": 254,
            "description": "Only under the me.allow_email_change setting; verification is reset"
          }
        },
        "additionalProperties": false
      },
      "MePasskey": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "credential_id": {
            "type": "string"
          },
          "alg": {
            "type": "integer"
          },
          "counter": {
            "type": "integer"
          },
          "transports": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "aaguid": {
            "type": [
              "string",
              "null"
            ]
          },
          "backup_eligible": {
            "type": "boolean"
          },
          "backed_up": {
            "type": "boolean"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_via": {
            "type": "string",
            "enum": [
              "interaction",
              "me",
              "recovery"
            ]
          },
          "created_at": {
            "type": "integer"
          },
          "last_used_at": {
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "credential_id",
          "alg",
          "counter",
          "transports",
          "aaguid",
          "backup_eligible",
          "backed_up",
          "name",
          "created_via",
          "created_at",
          "last_used_at"
        ]
      },
      "MePasskeyRegister": {
        "type": "object",
        "properties": {
          "response": {
            "type": "object",
            "properties": {},
            "additionalProperties": {},
            "description": "RegistrationResponseJSON from the browser"
          },
          "name": {
            "type": "string",
            "maxLength": 256
          }
        },
        "required": [
          "response"
        ],
        "additionalProperties": false
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/me": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "The signed-in person's profile (§8)",
        "security": [
          {
            "accountToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeProfile"
                }
              }
            }
          },
          "401": {
            "description": "invalid_token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "insufficient_scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "rate_limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "temporarily_unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "me"
        ],
        "summary": "Update display_name, and email when allowed (§8)",
        "security": [
          {
            "accountToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeProfilePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeProfile"
                }
              }
            }
          },
          "400": {
            "description": "invalid_request or email_invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "invalid_token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "insufficient_scope or email_change_not_allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "rate_limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "temporarily_unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/passkeys": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "The person's passkeys (§8)",
        "security": [
          {
            "accountToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Passkeys",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MePasskey"
                      }
                    }
                  },
                  "required": [
                    "items"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "invalid_token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "insufficient_scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "rate_limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "temporarily_unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "me"
        ],
        "summary": "Register a passkey (§8, TIO-ME-003)",
        "security": [
          {
            "accountToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MePasskeyRegister"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MePasskey"
                }
              }
            }
          },
          "400": {
            "description": "invalid_request or passkey_not_discoverable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "invalid_token or passkey_verification_failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "insufficient_scope, reauthentication_required or passkey_limit_reached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "rate_limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "temporarily_unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/passkeys/options": {
      "post": {
        "tags": [
          "me"
        ],
        "summary": "Registration options for a new passkey (§8, TIO-ME-003)",
        "security": [
          {
            "accountToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Options",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "publicKey": {
                      "type": "object",
                      "properties": {},
                      "additionalProperties": {},
                      "description": "PublicKeyCredentialCreationOptionsJSON for navigator.credentials.create()"
                    }
                  },
                  "required": [
                    "publicKey"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "invalid_token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "insufficient_scope or reauthentication_required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "rate_limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "temporarily_unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}